#2193 extract holdings chart code for swt/browser to own classes
