+++ name.abuchen.portfolio.ui/src/name/abuchen/portfolio/ui/views/charts/IPieChart.java
@@ -1,0 +1,13 @@
+package name.abuchen.portfolio.ui.views.charts;
+
+import org.eclipse.swt.widgets.Composite;
+import org.eclipse.swt.widgets.Control;
+
+import name.abuchen.portfolio.snapshot.ClientSnapshot;
+
+public interface IPieChart
+{
+    Control createControl(Composite parent);
+
+    void refresh(ClientSnapshot snapshot);
+}
